different thresholds

+++ Main.py
@@ -189,7 +189,7 @@ def train_model(unet, optimizer, scheduler, dataloader, dataset_sizes, device, l
                         scheduler.step()
 
                     epoch_loss = running_loss / dataset_sizes[phase]
-                    epoch_acc = running_corrects.double() / dataset_sizes[phase]
+                    epoch_acc = running_corrects / dataset_sizes[phase]
 
                     epoch_losses[phase].append(epoch_loss)
                     epoch_accuracies[phase].append(epoch_acc.item())
